doc: update op enabling doc (#24)
